Type:Bm Descr:Vérifier que le fichier existe bien.
git-svn-id: https://subversion.renater.fr/lodel/trunk@5600 9f73b02b-de1c-0410-bbc7-d36e22b4c400

+++ lodel/scripts/func.php
@@ -1451,6 +1451,8 @@ function thumbnail($path, $width = null, $height = null)
 	
 	$image_infos = pathinfo($path);
 
+	if(!file_exists($path)) return $path;
+
 	$tmp_path = tempnam(C::get('cacheDir', 'cfg'), 'thumb');
 	file_put_contents($tmp_path, file_get_contents($path));
